Completed Simulation. Now with Earth and Moon

+++ NewtonGravitySim/src/Space.h
@@ -10,7 +10,7 @@ class Space{
 public:
 	Space();
 	void addObject(SphericalObject& sphericalObject);
-	void addObject(dt::Vector2D position, double radius, double density = 1.0, dt::Vector2D velocity = dt::Vector2D(0.0, 0.0));
+	void addObject(dt::Vector2D position, long double radius, long double density = 1.0, dt::Vector2D velocity = dt::Vector2D(0.0, 0.0));
 	void update();
 	void draw(sf::RenderWindow&);
 };
