+++ NewtonGravitySim/src/SphericalObject.h
@@ -10,7 +10,7 @@ class SphericalObject{
 	long double radius, density;
 	sf::Color color;
 public:
-	SphericalObject(dt::Vector2D position, double radius, double density = 1.0, dt::Vector2D velocity = dt::Vector2D(0.0, 0.0));
+	SphericalObject(dt::Vector2D position, long double radius, long double density = 1.0, dt::Vector2D velocity = dt::Vector2D(0.0, 0.0));
 	long double getMass();
 	dt::Vector2D getPosition();
 	void update(long double& deltaTime, std::vector<SphericalObject>& otherSphericalObjects);
